// src/app/models/master.ts
export class ChangePassword {
    UserID: string;
    UserName: string;
    CurrentPassword: string;
    NewPassword: string;
}
export class EMailModel {
    EmailAddress: string;
    PortalAddress: string;
}
export class ForgotPassword {
    UserID: string;
    EmailAddress: string;
    NewPassword: string;
    Token: string;
}
export class UserNotification {
    ID: number;
    UserID: string;
    Message: string;
    HasSeen: boolean;
    CreatedOn: Date;
    ModifiedOn?: Date;
}
export class SendMail {
    Username: string;
    toEmail: string;
    otp: string;
}

export class LoginModel {
    UserName: string;
    Password: string;
    clientId: string;
}